Stop double-escaping URI segments for S3
An update to this module addressed the need for the URI segments within
a request to be double-encoded.  However, S3 is an exception
to this and double-encoding the URI segments will cause
the signature to be invalid.

From Step 2 in "creating a canonical request" in the following document:
https://docs.aws.amazon.com/general/latest/gr/sigv4-create-canonical-request.html

"Normalize URI paths according to RFC 3986. Remove redundant and
relative path components. Each path segment must be URI-encoded twice
(except for Amazon S3 which only gets URI-encoded once)."